# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: DataFlux API
  description: |
    DataFlux ist eine Universal AI-native Database für Medieninhalte mit einer Plugin-Architektur und Multi-Modal Search Capabilities.
    
    ## Features
    - **Multi-Modal Search**: Suche über Video, Audio, Bilder und Dokumente
    - **AI-Powered Analysis**: Automatische Inhaltsanalyse mit Machine Learning
    - **Real-time Processing**: Streaming und Batch-Verarbeitung
    - **Scalable Architecture**: Microservices-basierte Architektur
    - **Enterprise Security**: JWT Authentication und RBAC
    
    ## Authentication
    Alle API-Endpunkte (außer `/health` und `/docs`) erfordern Authentifizierung über JWT-Token.
    
    ```bash
    # Login
    curl -X POST http://localhost:8006/auth/login \
      -H "Content-Type: application/json" \
      -d '{"username":"admin","password":"admin123"}'
    
    # Use token in requests
    curl -H "Authorization: Bearer <token>" \
      http://localhost:8003/api/v1/search
    ```
  version: 1.0.0
  contact:
    name: DataFlux Support
    email: support@dataflux.com
    url: https://dataflux.com/support
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://api.dataflux.com
    description: Production server
...
paths:
...
  /api/v1/assets:
    post:
      tags: [Assets]
      summary: Upload asset
      description: Upload a new media asset for processing
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: Media file to upload
                collection_id:
                  type: string
                  format: uuid
                  description: Collection ID to add asset to
                priority:
                  type: string
                  enum: [low, normal, high]
                  default: normal
                  description: Processing priority
                metadata:
                  type: string
                  description: Additional metadata as JSON string
              required: [file]
      responses:
        '200':
          description: Asset uploaded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssetResponse'
        '400':
          description: Invalid file or validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    get:
      tags: [Assets]
      summary: List assets
      description: Get paginated list of assets
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
            minimum: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
        - name: collection_id
          in: query
          schema:
            type: string
            format: uuid
        - name: mime_type
          in: query
          schema:
            type: string
            example: video/mp4
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, processing, completed, failed]
      responses:
        '200':
          description: List of assets
          content:
            application/json:
              schema:
                type: object
                properties:
                  assets:
                    type: array
                    items:
                      $ref: '#/components/schemas/AssetResponse'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
  /api/v1/assets/{asset_id}:
    get:
      tags: [Assets]
      summary: Get asset details
      description: Get detailed information about a specific asset
      security:
        - bearerAuth: []
      parameters:
        - name: asset_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Asset details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssetResponse'
        '404':
          description: Asset not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # Search endpoints
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    AssetResponse:
      type: object
      properties:
        asset_id:
          type: string
          format: uuid
        file_name:
          type: string
        file_size:
          type: integer
          format: int64
        mime_type:
          type: string
        file_hash:
          type: string
        status:
          type: string
          enum: [pending, processing, completed, failed]
        collection_id:
          type: string
          format: uuid
        metadata:
          type: object
        thumbnail_path:
          type: string
        proxy_path:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        pages:
          type: integer
...
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
        timestamp:
          type: string
          format: date-time
...